pyproject.toml
[tool.poetry]
name = "velero-ui"
version = "0.4.1"
description = "Velero UI"
authors = ["Majid"]

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.0"
Flask-Cors = "^3.0.10"
Flask-RESTful = "^0.3.10"
bcrypt = "^3.2.0"
kubernetes = "^24.2.0"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
boto3 = "^1.34.14"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
